Revert "show user avatar in menu and chat userlist"
