--- app/components/site-settings.hbs
@@ -1,9 +1,0 @@
-<ThemeSelector
-  class="text-white"
-  @setTheme={{this.setTheme}}
-/>
-<LocaleSelector
-  class="text-white mt-1"
-  @selectedLocale={{this.intl.locale.value}}
-  @setLocale={{this.setLocale}}
-/>
