+++ app/components/sp-nav.js
@@ -1,10 +1,3 @@
 import Component from '@glimmer/component';
-import { inject as service } from '@ember/service';
 
-export default class SpNav extends Component {
-  @service
-  session;
-
-  @service
-  currentUser;
-}
+export default class SpNav extends Component {}
