+++ app/components/user-menu.hbs
@@ -1,5 +1,5 @@
 {{#if @showing}}
-  <ul role="menu" class="absolute right-0 classic:bg-df-pink blm:bg-black trans:bg-df-blue font-semibold z-10" style="top: 80px;">
+  <ul role="menu" class="absolute right-0 classic:bg-df-pink blm:bg-black trans:bg-df-blue font-semibold z-10" style="top: 85px;">
     <li class="hover:bg-df-blue rounded px-2 py-1">
       <LinkTo @route="home.user.settings">{{t "usermenu.settings"}}</LinkTo>
     </li>
