+++ app/controllers/home.js
@@ -13,9 +13,6 @@ export default class HomeController extends Controller {
   intl;
 
   @service
-  theme;
-
-  @service
   router;
 
   @service
@@ -54,7 +51,12 @@ export default class HomeController extends Controller {
 
   @action
   setTheme(theme) {
-    this.theme.setTheme(theme);
+    let element = document.getElementsByTagName('html')[0];
+    let currentTheme = `theme-${localStorage.getItem('datafruits-theme') || 'classic'}`;
+    element.classList.remove(currentTheme);
+    let themeName = `theme-${theme}`;
+    element.classList.add(themeName);
+    localStorage.setItem('datafruits-theme', theme);
   }
 
   @action
