+++ app/controllers/home/chat.js
@@ -26,8 +26,7 @@ export default class ChatController extends Controller {
       .authenticate('authenticator:devise', nick, pass)
       .then(() => {
         const token = this.session.data.authenticated.token;
-        const avatarUrl = this.currentUser.user.avatarUrl;
-        this.chat.push('authorize_token', { user: nick, timestamp: Date.now(), token: token, avatarUrl: avatarUrl });
+        this.chat.push('authorize_token', { user: nick, timestamp: Date.now(), token: token });
         return true;
       })
       .catch((/*reason*/) => {
