+++ app/routes/home/chat.js
@@ -15,9 +15,8 @@ export default class ChatRoute extends Route {
         .load()
         .then(() => {
           const nick = this.currentUser.user.username;
-          const avatarUrl = this.currentUser.user.avatarUrl;
           const token = this.session.data.authenticated.token;
-          this.chat.push('authorize_token', { user: nick, timestamp: Date.now(), token: token, avatarUrl: avatarUrl });
+          this.chat.push('authorize_token', { user: nick, timestamp: Date.now(), token: token });
         })
         .catch(() => this.session.invalidate());
     }
