--- app/routes/home/site-settings.js
@@ -1,3 +1,0 @@
-import Route from '@ember/routing/route';
-
-export default class HomeSiteSettingsRoute extends Route {}
