+++ app/templates/components/datafruits-chat.hbs
@@ -17,9 +17,9 @@
     </div>
   {{/if}}
   <ul id="userlist" class="overflow-y-scroll flex-grow hidden md:block md:w-1/5 pl-4">
-    {{#each-in this.joinedUsers as |username user|}}
-      <UserListItem @user={{user}} />
-    {{/each-in}}
+    {{#each joinedUsers as |joinedUser|}}
+      <UserListItem @user={{joinedUser}} />
+    {{/each}}
   </ul>
 </div>
 <div class="flex flex-col justify-center items-center w-full">
