+++ app/templates/components/sp-nav.hbs
@@ -101,39 +101,20 @@
             {{t "subscribe.title"}}
           </LinkTo>
         </li>
-        <li class="py-2 pr-3 w-full border-dashed border-white border-t-4">
-          <LinkTo
-            class="text-white text-shadow-light"
-            aria-label={{t "settings.title"}}
-            @route="home.site-settings">
-            {{t "settings.title"}}
-          </LinkTo>
-        </li>
       </ul>
     {{/if}}
   </div>
   {{#if (not @menuOpen)}}
     <DjDonateButton class="block md:hidden ml-5" />
     <div class="">
-      {{#if this.session.isAuthenticated}}
-        <a href="#" class="" role="button" {{on "click" this.args.toggleUserMenu}}>
-          <img
-            class="inline rounded-lg"
-            style="height: 3rem;"
-            src="{{this.currentUser.user.avatarUrl}}"
-            align="center"
-            alt="{{ t "chat.user_avatar" }}" / >
-        </a>
-      {{else}}
-        <LinkTo @route="home"
-          @aria-label={{t "blog"}}>
-          <img
-            class=""
-            id="circle-logo"
-            src="/assets/images/logo.png"
-            alt={{t "site_title"}}>
-        </LinkTo>
-      {{/if}}
+      <LinkTo @route="home"
+        @aria-label={{t "blog"}}>
+        <img
+          class=""
+          id="circle-logo"
+          src="/assets/images/logo.png"
+          alt={{t "site_title"}}>
+      </LinkTo>
     </div>
   {{/if}}
 </nav>
