--- app/templates/home/site-settings.hbs
@@ -1,6 +1,0 @@
-<div class="mt-5 pt-2 px-5">
-  {{page-title (t "settings.title")}}
-  <h1>{{t "settings.title"}}</h1>
-
-  <SiteSettings />
-</div>
