+++ tests/integration/components/user-list-item-test.js
@@ -9,9 +9,8 @@ module('Integration | Component | user list item', function (hooks) {
   test('it renders', async function (assert) {
     // Set any properties with this.set('myProperty', 'value');
     // Handle any actions with this.on('myAction', function(val) { ... });
-    this.set('user', { metas: [{ avatarUrl: 'garfield.gif', username: 'garfield' }] });
 
-    await render(hbs`<UserListItem @user={{user}} />`);
+    await render(hbs`{{user-list-item}}`);
 
     assert.dom('*').hasText('');
   });
