--- tests/unit/services/theme-test.js
@@ -1,12 +1,0 @@
-import { module, test } from 'qunit';
-import { setupTest } from 'ember-qunit';
-
-module('Unit | Service | theme', function (hooks) {
-  setupTest(hooks);
-
-  // TODO: Replace this with your real tests.
-  test('it exists', function (assert) {
-    let service = this.owner.lookup('service:theme');
-    assert.ok(service);
-  });
-});
