+++ translations/ja.json
@@ -1,7 +1,4 @@
 {
-  "settings": {
-    "title": "サイト設定"
-  },
   "home": {
     "upcoming_shows": "近日放送予定",
     "latest_podcasts": "最新のアーカイブ",
@@ -116,7 +113,7 @@
       "now_playing": "録音を再生しています"
     }
   },
-  "locale": "言語",
+  "locale": "ロケール",
   "theme": { "label": "テーマ" },
   "donate": "寄付する"
 }
